Move multimodal filters to interleaved
